refactor(event): remove AuditChannelPermissionUpdate
i honestly don't know why i added this in the first place. it seems out-of-place given that AuditRoleUpdate provides permission updates without a special audit type. i probably saw that there was a special function in the python version called `on_guild_channel_perms_update` and decided to add it as an audit type.